prob send/rcv;TreeSet instead of PriQ

+++ src/edu/purdue/cs505/SenderThread.java
@@ -8,6 +8,7 @@ import java.net.InetAddress;
 import java.net.SocketException;
 import java.net.UnknownHostException;
 import java.util.Iterator;
+import java.util.Random;
 
 class SenderThread extends Thread {
   private RChannel rChannel;
@@ -36,8 +37,9 @@ class SenderThread extends Thread {
             }
 
             itr = rChannel.sendBuffer.iterator();
-            int temp = rChannel.sendBuffer.peek().getSeqNo();
-            // Debugger.print(2, "Sending: " + temp + " to: " + (temp + 31));
+            short temp = rChannel.sendBuffer.peek().getSeqNo();
+           // if(temp >= 32600 || temp <= 32)
+             // Debugger.print(2, "Sending: " + temp + " to: " + (temp + 31)%Short.MAX_VALUE);
             for (int sendCount = 0; sendCount < RChannel.bufferLength
                 && itr.hasNext(); sendCount++) {
               ByteArrayOutputStream baos = new ByteArrayOutputStream();
@@ -50,7 +52,13 @@ class SenderThread extends Thread {
                   InetAddress.getByName("localhost"),
                   rChannel.getDestinationPort());
               Debugger.print(1, "UDP Send " + m.toString());
-              rChannel.getUdpChannel().send(out);
+              //Test:Create Congestion - Simulate missing frames.
+//              Random generator = new Random(); 
+//              int i = generator.nextInt(2); //a random number either 0/1.
+//              if(i%2 == 0)
+//              {
+                rChannel.getUdpChannel().send(out);
+//              }
               m.incResndCount();
             }
           } else {
